backend/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
celery
redis
pymysql